made everything integers

+++ Calculate.Tests/CalculatorTests.cs
@@ -10,9 +10,9 @@ public class CalculatorTests
     [InlineData(9, 6, 3)]
     [InlineData(-3, -5, 2)]
     [InlineData(5.5, 2.2, 3.3)]
-    public void Add_ReturnSumOfTwoValues_Equal(double expected, double val1, double val2)
+    public void Add_ReturnSumOfTwoValues_Equal(int expected, int val1, int val2)
     {
-        double sum = Calculator.Add(val1, val2);
+        int sum = Calculator.Add(val1, val2);
         Assert.Equal(expected, sum);
     }
 
@@ -21,9 +21,9 @@ public class CalculatorTests
     [InlineData(3, 6, 3)]
     [InlineData(-7, -5, 2)]
     [InlineData(7.8, 8, 0.2)]
-    public void Subtract_ReturnDifferenceOfTwoValues_Equal(double expected, double val1, double val2)
+    public void Subtract_ReturnDifferenceOfTwoValues_Equal(int expected, int val1, int val2)
     {
-        double difference = Calculator.Subtract(val1, val2);
+        int difference = Calculator.Subtract(val1, val2);
         Assert.Equal(expected, difference);
     }
 
@@ -31,9 +31,9 @@ public class CalculatorTests
     [InlineData(4, 2, 2)]
     [InlineData(18, 6, 3)]
     [InlineData(6.6, 3.3, 2)]
-    public void Multiply_ReturnProductOfTwoValues_Equal(double expected, double val1, double val2)
+    public void Multiply_ReturnProductOfTwoValues_Equal(int expected, int val1, int val2)
     {
-        double product = Calculator.Multiple(val1, val2);
+        int product = Calculator.Multiple(val1, val2);
         Assert.Equal(expected, product);
     }
 
@@ -41,9 +41,9 @@ public class CalculatorTests
     [InlineData(1, 2, 2)]
     [InlineData(2, 6, 3)]
     [InlineData(2.5, 5, 2)]
-    public void Divide_ReturnQuotientOfTwoValues_Equal(double expected, double val1, double val2)
+    public void Divide_ReturnQuotientOfTwoValues_Equal(int expected, int val1, int val2)
     {
-        double quotient = Calculator.Divide(val1, val2);
+        int quotient = Calculator.Divide(val1, val2);
         Assert.Equal(expected, quotient);
     }
 
@@ -62,10 +62,10 @@ public class CalculatorTests
     [InlineData('-', -5.5, 36.5, 42)]
     [InlineData('*', 19.2, 9.6, 2)]
     [InlineData('/', 1.5, -18, -12)]
-    public void MathematicalOperations_VariousOperations_Equal(char operatorSymbol, double expected, double val1, double val2)
+    public void MathematicalOperations_VariousOperations_Equal(char operatorSymbol, int expected, int val1, int val2)
     {
         var operation = Calculator.MathematicalOperations[operatorSymbol];
-        double result = operation(val1, val2);
+        int result = operation(val1, val2);
         Assert.Equal(expected, result);
     }
 
@@ -73,7 +73,7 @@ public class CalculatorTests
     public void TryCalculate_TooManyParams_FalseAndNullAnswer()
     {
         string input = "7 + 8 6";
-        Assert.False(Calculator.TryCalculate(input, out double? ans));
+        Assert.False(Calculator.TryCalculate(input, out int? ans));
         Assert.Null(ans);
     }
 
@@ -81,7 +81,7 @@ public class CalculatorTests
     public void TryCalculate_BadOperatorSymbol_FalseAndNullAnswer()
     {
         string input = "3 , 4";
-        Assert.False(Calculator.TryCalculate(input, out double? ans));
+        Assert.False(Calculator.TryCalculate(input, out int? ans));
         Assert.Null(ans);
     }
 }
